requirements.txt
pandas>=2.2
openpyxl>=3.1
PyYAML>=6.0